Many bug fixes and output changes.
Many of the bugs were related to marking of trajectories as
periodic. This resulted in changes in CGMap and pickling of
ForceMatch. There were bugs in loading forces as well in
CGMap. Caching of dimension calculation was added to some Forces (huge
performance gain). Should be added to the rest. There was a bug in the
neighborlist image correction that was fixed in Util.c, but not
there. Also removed extraneous code in Util. Extra debugging
code/assertions were added in some places. The most impactful on
performance may be in the Mesh, but is worth it for now. A better
definition of error was added to ForceMatch which doesn't scale as the
number of particles increase. The test scripts were updeated to
reflect theses changes.